exclude test coverage for gcp_metadata
